pyproject.toml
[project]
name = "python-rag-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "black>=26.3.1",
    "chromadb>=1.5.5",
    "isort>=8.0.1",
    "langchain>=1.2.13",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.12",
    "langchain-prompty>=1.0.0",
    "langchain-tavily>=0.2.17",
    "langchain-text-splitters>=1.1.1",
    "openai>=2.29.0",
    "pypdf>=6.9.2",
    "python-dotenv>=1.2.2",
    "tavily>=1.1.0",
]
